cmd/snaplock/runinhibit: save revision when snap is inhibited
This is handy to be used by "snap run" to get snap info when
the "current" symlink is not available (i.e. during a refresh
after the snap is unlinked).

An extra data JSON file is created <inhibit-dir>/<snap>.<hint>. This is
needed to carry more information that is needed by "snap run" while
the snap is inhibited. It can be expanded to add more data as needed.

Signed-off-by: Zeyad Gouda <zeyad.gouda@canonical.com>